pyproject.toml
[project]
name = "visitaspro"
version = "0.1.0"
description = "Sistema de visitas corporativas"
authors = [
    {name = "VitorJonatasPB", email = "vitor.jonatas98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==6.0.2",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.0",
    "django-cors-headers>=4.0",
    "asgiref==3.11.1",
    "diff-match-patch==20241021",
    "django-admin-filters==0.1.3",
    "django-admin-list-filter-dropdown==1.0.3",
    "django-autocomplete-light==3.12.1",
    "django-filter==25.2",
    "django-import-export==4.4.0",
    "django-jazzmin==3.0.2",
    "django-js-asset==3.1.2",
    "django-lineup==1.0.0",
    "django-mptt==0.18.0",
    "django-smart-selects==1.7.2",
    "et-xmlfile==2.0.0",
    "numpy==2.4.3",
    "openpyxl==3.1.5",
    "pandas==3.0.1",
    "pillow==12.1.1",
    "psycopg==3.3.2",
    "psycopg-binary==3.3.2",
    "python-dateutil==2.9.0.post0",
    "six==1.17.0",
    "sqlparse==0.5.5",
    "tablib==3.9.0",
    "tzdata==2025.3",
    "gunicorn>=25.3.0,<26.0.0",
    "whitenoise>=6.12.0,<7.0.0",
    "dj-database-url>=3.1.2,<4.0.0",
    "python-dotenv>=1.2.2,<2.0.0",
    "requests>=2.33.1,<3.0.0",
    "reportlab (>=4.5.0) ; python_version >= \"3.12\" and python_version < \"4\"",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
